requirements.txt
celery==5.3.6
Paver==1.3.4
pandas==2.2.0
Unidecode==1.3.8
flake8==7.0.0
fastapi==0.109.2
fastapi-pagination==0.12.15
uvicorn==0.27.1
redis==5.0.1
requests==2.31.0
PyMySQL==1.1.0
SQLAlchemy==2.0.25
python-decouple==3.8
strawberry-graphql==0.208.0